separated the packages since pdf has requirements not needed by simple text version
