+++ claim_extractor/src/claim_extractor/schemas/__init__.py
@@ -1,0 +1,15 @@
+from .loader import (
+    load_raw_schema,
+    LINKED_CLAIM,
+    OPEN_CRED,
+    LINKED_TRUST,
+    CLAIM_SCHEMAS
+)
+
+__all__ = [
+    'load_raw_schema',
+    'LINKED_CLAIM',
+    'OPEN_CRED',
+    'LINKED_TRUST',
+    'CLAIM_SCHEMAS'
+]
