+++ pdf_parser/README.md
@@ -1,1 +1,8 @@
 # Parse PDF files into verifiable claims
+
+## Local development/testing
+
+```
+pip install -r requirements.txt
+pip install -e .
+```
