--- pyproject.toml
@@ -1,50 +1,0 @@
-[build-system]
-requires = ["hatchling"]
-build-backend = "hatchling.build"
-
-[project]
-name = "linked-claims-extractor"
-version = "0.1.1"
-
-description = "Extract structured claims from text and PDFs"
-readme = "README.md"
-authors = [
-    { name = "Golda Velez", email = "gvelez17@gmail.com" }
-]
-license = { text = "MIT" }
-requires-python = ">=3.8"
-classifiers = [
-    "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: MIT License",
-    "Operating System :: OS Independent",
-]
-dependencies = [
-    "langchain-anthropic>=0.2.4",
-    "langchain>=0.1.0",
-    "langchain-community>=0.0.10",
-    "openai>=1.0.0",
-    "requests>=2.25.0",
-    "python-dotenv>=1.0.0",
-    "pytest>=7.0.0",
-    "pypdf>=3.0.0",
-    "pdfminer.six>=20221105",
-]
-
-urls.homepage = "https://github.com/Cooperation-org/linked-claims-extractor"
-urls.bug_tracker = "https://github.com/Cooperation-org/linked-claims-extractor/issues"
-
-
-[tool.hatch.build.targets.wheel]
-packages = [
-    "src/claim_extractor",
-    "src/pdf_parser",
-]
-
-[tool.pytest.ini_options]
-markers = [
-    "integration: tests that use any real API",
-    "slow: marks tests as slow",
-]
-
-[tool.setuptools]
-package-dir = {"" = "src"}
